// desktop/src/open-api/model/permissionScope.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Whether a permission applies app-wide or within a single group.
 */
export type PermissionScope = 'APP' | 'GROUP';

export const PermissionScope = {
    App: 'APP' as PermissionScope,
    Group: 'GROUP' as PermissionScope
};